new "checker" program/lib that uses hfst/cg/suggest libs as Pipeline
Use configure option --enable-checker to build it; not built by
default since it requires libarchive and a patched version of vislcg3.
People who only work on language data probably only need
divvun-suggest.

New libraries libdivvun-suggest and libdivvun-checker (we might want
to merge those later).

The Pipeline object in checker.cpp runs lots of PipeCmd objects as a
fake (single-process) pipeline. It can be started from a plain
pipespec.xml config with paths to language data, or from a zipped
archive containing the pipespec.xml (has to be named that!) with
references to other files in the archive. There is a DTD documenting
the pipespec format, and divvun-validate-checker script for validating
pipespec.xml files.

The new divvun-checker binary is an example usage of
libdivvun-checker, for easy testing of pipespec.xml's and archives.

The pipespec can define multiple, named pipelines, and some language
metadata. It also contains a <prefs> element for preferences; actually
using those and letting the user select them is still TODO.

Tests and Travis build script updated to work with these new features.
Travis does some caching now.

This is a squashed commit of
https://github.com/divvun/divvun-gramcheck/tree/librarisation-zip-reading