divvun-blanktag: analyse blank-wf-blank-seqs; tag matched readings
divvun-blanktag takes an FST as argument, reads CG input and uses the
FST to add readings to cohorts that match on the wf surrounded by
blanks.

So this can be used to e.g. add error tags that are dependent on
spaces before/after, or tag the first word after a linebreak, etc.