allow words in between R:LEFT id/target, support &ADDED-BEFORE-BLANK
Put the tag &ADDED-BEFORE-BLANK on a reading of an added cohort in
order to get it as part of a suggestion where you want the form of
this cohort to appear before the blank at that position. Useful
for e.g. adding a comma before a word.

R:RIGHT still TODO, also should put this huge lambda into its own
function