Added beam and maxweight
This is done to reduce the number of (mostly irrelevant) suggestions, and remove unwanted suggestions (those with very high weights). This should make the speller a bit faster, and make the life easier for the CG rules that are applied after the speller. The exact beam value needs a lot of testing, the present value (15) is chosen to allow dynamic compounds with only one dynamic border (=10) + a small amount of additional changes. Let’s see how this works out.