Combine SUGGESTWF with other suggestions, don't assume it covers full squiggle
Lets us combine typos into complex errors.

This is a breaking change – rules like

    COPY (VSTR:"<$2\ $1>" SUGGESTWF) TARGET ("<(.*)>"r &missing-space-before) IF (-1 ("<(.*)>"r))

need to be changed to

    COPY   (VSTR:"<\ $1>"S SUGGESTWF) TARGET ("<(.*)>"r &missing-space-before)

(ie. don't include the preceding cohort in the suggestion)

There is a new tag DROP-PRE-BLANK for the case where you want to
remove the blank before a suggestion, e.g.

    COPY (DROP-PRE-BLANK) TARGET (&space-before-punct-mark)

"<ovddasvástadus>"
	"ovddasvástádus" Err/Orth-a-á N <BE-Ill-Any> Sem/Perc-emo Sg Nom <W:0.0> @<SUBJ &space-before-punct-mark #6->6 ID:6
:
"<.>"
	"." CLB <W:0.0> <SpaceBeforePunctMark> &space-before-punct-mark DROP-PRE-BLANK #7->7 ID:7 R:LEFT:6
	"." CLB <W:0.0> <SpaceBeforePunctMark> &space-before-punct-mark DROP-PRE-BLANK #7->7 ID:7 R:LEFT:6